CRM/backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Mail==0.9.1
python-dotenv==1.0.1
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==1.10.15
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-dotenv==1.0.1
alembic==1.13.2
jinja2==3.1.4

